Start replacing the monotone decomposer with a monotone triangulator
The monotone decomposer was too complicated, hard to debug, and did not have
good abstractions internally. So this commit starts making a triangulator
instead. In a later commit, the monotone decomposer will be rebuilt as a layer
on top of the triangulator, by merging the output triangles into larger, convex
polygons.