Use edges_ inside of CreateSplitChildren
Use edges_ directly inside of CreateSplitChildren, instead of going through the
edge accessor. The edge accessor returns a const&, which prevents MakeForward
from converting it into a rvalue reference.

Also add a check to verify that const references are not being passed to
MakeForward when the SplitKey was created from an rvalue reference.