Make the PluckerLine equality sensitive to orientation
Some of the other test cases are subtly relying on the PluckerLine orientation.
Notably reversing the orientation of all PluckerLines inside the constructor
causes tests using ConvexPolygon to fail. Since the other code is relying on
the orientation, the orientation should be taken into account when considering
whether two lines are equal.