Use the split ranges instead of the split info in ConvexVertexAabbTracker
ConvexPolygonSplitInfo can contain an actual line and points. So eventually
copying it could involve heap allocations. The type is also templated. On the
other hand, ConvexPolygonSplitRanges contains only integral types with no heap
allocations. Since ConvexVertexAabbTracker is only using the ranges portion of
the split info, it makes sense to pass in only the ranges.