Modify ConnectedEdge to have a back pointer to its polygon
Previously ConnectedEdge had a pointer to its neighbor polygon and the index of
its neighbor edge. Now it instead contains a pointer back to its own polygon
and a pointer to its neighbor edge.

If the edges are connected 1 to 1, the ConnectedEdges themselves take the same
amount of space in both approaches. One difference with the second approach is
that from just a pointer to a ConnectedEdge, both edges and both polygons can
be recovered. So externally the second approach takes less memory. Another
difference is that the second approach takes less memory if it has extra
neighbors, because those extra neighbors are only an edge pointer.