Fix a bug in PlanePartitioner
The PlanePartitioner was assuming that the edges that need to be split would
have null coincident edge info. However, the polygons produced by the
BSPTraverser actually never have a null coincident edge, because they are
coincident with at least the last split.

So instead look for edges whose first coincident split matches the last
coincident split, meaning there is only one coincident split for the edge.