Fix a bug with merging edges in a border polygon
If a border polygon is split multiple times after it became a border polygon,
then the edge_first_coincident field in its edges does not provide enough
information about which order to merge the edges in. So sort the edges using
the created_by field before trying to merge them.